+++ R/admiral-package.R
@@ -4,7 +4,7 @@
 #' @importFrom dplyr across arrange bind_rows case_when desc ends_with
 #'   everything filter full_join group_by if_else mutate n pull rename
 #'   rename_with row_number select slice semi_join starts_with transmute ungroup
-#'   vars n_distinct union distinct summarise coalesce bind_cols na_if tibble
+#'   n_distinct union distinct summarise coalesce bind_cols na_if tibble
 #' @importFrom magrittr %>%
 #' @importFrom rlang := abort arg_match as_function as_label as_name as_string
 #'   call2 caller_env call_name current_env .data enexpr eval_bare eval_tidy